rename RenewConventionMagicLink to RenewExpiredJwt and move to auth router
