+++ back/src/adapters/primary/routers/convention/convention.e2e.test.ts
@@ -19,8 +19,6 @@ import {
   expectHttpResponseToEqual,
   expectToEqual,
   expiredMagicLinkErrorMessage,
-  frontRoutes,
-  type Role,
   type TechnicalRoutes,
   technicalRoutes,
   type UnauthenticatedConventionRoutes,
@@ -31,13 +29,10 @@ import {
 import type { HttpClient } from "shared-routes";
 import { createSupertestSharedClient } from "shared-routes/supertest";
 import { match } from "ts-pattern";
-import type { AppConfig } from "../../../../config/bootstrap/appConfig";
 import type { BasicEventCrawler } from "../../../../domains/core/events/adapters/EventCrawlerImplementations";
-import {
-  type GenerateConnectedUserJwt,
-  type GenerateConventionJwt,
-  makeGenerateJwtES256,
-  makeVerifyJwtES256,
+import type {
+  GenerateConnectedUserJwt,
+  GenerateConventionJwt,
 } from "../../../../domains/core/jwt";
 import type { InMemoryUnitOfWork } from "../../../../domains/core/unit-of-work/adapters/createInMemoryUow";
 import { AppConfigBuilder } from "../../../../utils/AppConfigBuilder";
@@ -91,7 +86,6 @@ describe("convention e2e", () => {
   let inMemoryUow: InMemoryUnitOfWork;
   let eventCrawler: BasicEventCrawler;
   let gateways: InMemoryGateways;
-  let appConfig: AppConfig;
 
   beforeEach(async () => {
     const testApp = await buildTestApp(new AppConfigBuilder().build());
@@ -103,7 +97,6 @@ describe("convention e2e", () => {
       generateConventionJwt,
       generateConnectedUserJwt,
       inMemoryUow,
-      appConfig,
     } = testApp);
 
     unauthenticatedRequest = createSupertestSharedClient(
@@ -760,123 +753,6 @@ describe("convention e2e", () => {
           message: errors.convention.notFound({
             conventionId: unknownId,
           }).message,
-        },
-      });
-    });
-  });
-
-  describe(`${displayRouteName(
-    conventionMagicLinkRoutes.renewConvention,
-  )} renews a magic link`, () => {
-    it("200 - sends the updated magic link", async () => {
-      const validConvention = new ConventionDtoBuilder().build();
-
-      const agency = AgencyDtoBuilder.create(validConvention.agencyId)
-        .withName("TEST-name")
-        .withSignature("TEST-signature")
-        .build();
-
-      const convention = new ConventionDtoBuilder().build();
-      inMemoryUow.conventionRepository.setConventions([convention]);
-      inMemoryUow.agencyRepository.agencies = [toAgencyWithRights(agency)];
-
-      gateways.timeGateway.setNextDate(new Date());
-
-      generateConventionJwt = makeGenerateJwtES256<"convention">(
-        appConfig.jwtPrivateKey,
-        3600 * 24, // one day
-      );
-      const shortLinkIds = ["shortLink1", "shortLinkg2"];
-      gateways.shortLinkGenerator.addMoreShortLinkIds(shortLinkIds);
-
-      const originalUrl = `${appConfig.immersionFacileBaseUrl}/${frontRoutes.conventionToSign}`;
-
-      const expiredJwt = generateConventionJwt(
-        createConventionMagicLinkPayload({
-          id: validConvention.id,
-          role: "beneficiary",
-          email: validConvention.signatories.beneficiary.email,
-          now: new Date(),
-        }),
-      );
-
-      const response = await unauthenticatedRequest.renewMagicLink({
-        queryParams: {
-          expiredJwt,
-          originalUrl: encodeURIComponent(originalUrl),
-        },
-      });
-
-      expect(response.status).toBe(200);
-
-      await processEventsForEmailToBeSent(eventCrawler);
-
-      const sentEmails = gateways.notification.getSentEmails();
-
-      expect(sentEmails).toHaveLength(1);
-
-      const email = expectEmailOfType(sentEmails[0], "MAGIC_LINK_RENEWAL");
-      expect(email.recipients).toEqual([
-        validConvention.signatories.beneficiary.email,
-      ]);
-
-      const magicLink = await shortLinkRedirectToLinkWithValidation(
-        email.params.magicLink,
-        technicalRoutesClient,
-      );
-
-      const newUrlStart = `${originalUrl}?jwt=`;
-
-      expect(magicLink.startsWith(newUrlStart)).toBeTruthy();
-      const renewedJwt = magicLink.replace(newUrlStart, "");
-      expect(renewedJwt !== expiredJwt).toBeTruthy();
-      expect(
-        makeVerifyJwtES256(appConfig.jwtPublicKey)(renewedJwt),
-      ).toBeDefined();
-    });
-
-    it("400 - renew not allowed for back-office", async () => {
-      inMemoryUow.conventionRepository.setConventions([convention]);
-      inMemoryUow.agencyRepository.agencies = [
-        toAgencyWithRights(
-          AgencyDtoBuilder.create(convention.agencyId)
-            .withName("TEST-name")
-            .withSignature("TEST-signature")
-            .build(),
-        ),
-      ];
-      gateways.timeGateway.setNextDate(new Date());
-
-      generateConventionJwt = makeGenerateJwtES256<"convention">(
-        appConfig.jwtPrivateKey,
-        3600 * 24, // one day
-      );
-
-      const unsupportedRole: Role = "back-office";
-
-      const response = await unauthenticatedRequest.renewMagicLink({
-        queryParams: {
-          expiredJwt: generateConventionJwt(
-            createConventionMagicLinkPayload({
-              id: convention.id,
-              role: unsupportedRole as ConventionRole,
-              email: convention.establishmentTutor.email,
-              now: gateways.timeGateway.now(),
-            }),
-          ),
-          originalUrl: encodeURIComponent(
-            `https://${appConfig.immersionFacileBaseUrl}/${frontRoutes.assessment}`,
-          ),
-        },
-      });
-
-      expectHttpResponseToEqual(response, {
-        status: 400,
-        body: {
-          message: errors.convention.roleHasNoMagicLink({
-            role: unsupportedRole,
-          }).message,
-          status: 400,
         },
       });
     });
