+++ shared/src/routes/convention.routes.ts
@@ -15,7 +15,6 @@ import {
   markPartnersErroredConventionAsHandledRequestSchema,
   paginatedConventionReadSchema,
   renewConventionParamsSchema,
-  renewMagicLinkRequestSchema,
   renewMagicLinkResponseSchema,
   sendSignatureLinkRequestSchema,
   transferConventionToAgencyRequestSchema,
@@ -203,16 +202,7 @@ export const unauthenticatedConventionRoutes = defineRoutes({
     requestBodySchema: shareLinkByEmailSchema,
     responses: { 200: expressEmptyResponseBody, 400: httpErrorSchema },
   }),
-  renewMagicLink: defineRoute({
-    url: "/renew-magic-link",
-    method: "get",
-    queryParamsSchema: renewMagicLinkRequestSchema,
-    responses: {
-      200: expressEmptyResponseBody,
-      400: httpErrorSchema,
-      404: httpErrorSchema,
-    },
-  }),
+
   findSimilarConventions: defineRoute({
     url: "/find-similar-immersion",
     method: "get",
